// frontend/src/api/generated/models/ApartmentsStatisticsOverviewDTO.ts
export interface ApartmentsStatisticsOverviewDTO {
    /**
     * 
     * @type {number}
     * @memberof ApartmentsStatisticsOverviewDTO
     */
    totalApartments: number;
    /**
     * 
     * @type {number}
     * @memberof ApartmentsStatisticsOverviewDTO
     */
    totalRooms: number;
    /**
     * 
     * @type {number}
     * @memberof ApartmentsStatisticsOverviewDTO
     */
    totalOccupiedRooms: number;
    /**
     * 
     * @type {number}
     * @memberof ApartmentsStatisticsOverviewDTO
     */
    totalFreeRooms: number;
    /**
     * 
     * @type {number}
     * @memberof ApartmentsStatisticsOverviewDTO
     */
    overallOccupancyRatePercent: number;
    /**
     * 
     * @type {number}
     * @memberof ApartmentsStatisticsOverviewDTO
     */
    totalActiveMonthlyRent: number;
    /**
     * 
     * @type {number}
     * @memberof ApartmentsStatisticsOverviewDTO
     */
    totalIncomeCollectedAllTime: number;
    /**
     * 
     * @type {number}
     * @memberof ApartmentsStatisticsOverviewDTO
     */
    currentMonthIncomeCollected: number;
    /**
     * 
     * @type {number}
     * @memberof ApartmentsStatisticsOverviewDTO
     */
    totalCostsAllTime: number;
    /**
     * 
     * @type {number}
     * @memberof ApartmentsStatisticsOverviewDTO
     */
    currentMonthCosts: number;
    /**
     * 
     * @type {number}
     * @memberof ApartmentsStatisticsOverviewDTO
     */
    unassignedCostsAllTime: number;
    /**
     * 
     * @type {number}
     * @memberof ApartmentsStatisticsOverviewDTO
     */
    netResultCurrentMonth: number;
    /**
     * 
     * @type {number}
     * @memberof ApartmentsStatisticsOverviewDTO
     */
    netResultAllTime: number;
}